+++ Tutorials/02-Blazor/02D-ReduxDevToolsTutorial/ReduxDevToolsTutorial/Client/Program.cs
@@ -16,13 +16,15 @@ namespace FluxorBlazorWeb.ReduxDevToolsTutorial.Client
 
 			builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 
-			builder.Services.AddFluxor(o => o
-				.ScanAssemblies(typeof(Program).Assembly)
-				.UseReduxDevTools(rdt =>
+			builder.Services.AddFluxor(o =>
+			{
+				o.ScanAssemblies(typeof(Program).Assembly);
+#if DEBUG
+				o.UseReduxDevTools(rdt =>
 				{
 					rdt.Name = "Fluxor ReduxDevTools sample";
 
-					//Example of using Newtonsoft, and optionally providing serialization settings
+					// Example of using Newtonsoft, and optionally providing serialization settings
 					rdt.UseNewtonsoftJson(_ =>
 							new Newtonsoft.Json.JsonSerializerSettings
 							{
@@ -33,16 +35,19 @@ namespace FluxorBlazorWeb.ReduxDevToolsTutorial.Client
 							}
 					);
 
-					//Example of using System.Text.Json, and optionally providing serialization options
-					//Note: This won't work with Routing Middleware enabled
+					// Example of using System.Text.Json, and optionally providing serialization options
+					// Note: This won't work if RoutingMiddleware is enabled, because
+					// System.Text.Json doesn't deserialize classes without a parameterless constructor
+					// in .net 3.1
 					//rdt.UseSystemTextJson(_ =>
 					//	new System.Text.Json.JsonSerializerOptions
 					//	{
 					//		PropertyNamingPolicy = System.Text.Json.JsonNamingPolicy.CamelCase
 					//	}
 					//);
-				})
-			);
+				});
+#endif
+			});
 			await builder.Build().RunAsync();
 		}
 	}