+++ Docs/releases.md
@@ -1,5 +1,14 @@
 # Releases
 
+## New in 5.1
+ * Fixed `IStateSelection<TState, TValue>` bug that threw exception stating the selector has
+    already been set.  ([#252](https://github.com/mrpmorris/Fluxor/issues/252))
+ * Added an optional `Action<TValue> selectedValueChanged` to `IStateSelection<TState, TValue>.Select`
+    that is executed whenever the selected value changes. This is a convenient alternative to hooking up event handlers.
+ * Added `event EventHandler<TValue> SelectedValueChanged` to `IStateSelection<TState, TValue>` for strongly
+    typed event args when the selected value changes.
+ * Changed to avoid .net memory leak in dependency injection ([#271](https://github.com/mrpmorris/Fluxor/issues/271))
+
 ## New in 5.0
  * **Breaking change**: Removed need to reference `_content/Fluxor.Blazor.Web/scripts/index.js` ([#235](https://github.com/mrpmorris/Fluxor/issues/235))
  * **Breaking change**: Removed `IState<TState>` generic `StateChanged` event.